feat(zfb-diagnostics): extract Diagnostic, FramedError, render_framed into standalone crate
Promote the core diagnostic types and rendering logic from
crates/zfb/src/diagnostics.rs into a new zfb-diagnostics crate. This
lets the render and build crates emit framed errors without depending on
the top-level kitchen-sink zfb crate.

What moved to zfb-diagnostics:
- Diagnostic struct and impl (new, with_source, try_attach_source_from_disk)
- FramedError wrapper and Display/Error impls
- render_framed function
- project_relative helper
- locate_export_ident helper (now pub for cross-crate use)
- from_js_runtime_error_with_decoder (generic, accepts a decode-fn closure)
- DecodedPosition trait (decouples zfb-diagnostics from zfb-render)

What stays in crates/zfb/src/diagnostics.rs (depends on zfb-content/zfb-render):
- from_frontmatter_error / from_tsx_frontmatter_error
- from_directive_diagnostic
- from_paths_error / from_paths_extract_error
- from_js_runtime_error (thin wrapper that adapts zfb_render::sourcemap::DecodedFrame
  to the DecodedPosition trait and delegates to the generic function)

All items are re-exported from crates/zfb so existing callers using
zfb::diagnostics::* keep compiling unchanged.
